Counts the RGs automatically

+++ hybrid-labs-destroy.ps1
@@ -1,10 +1,7 @@
 # Assumes you created the environment using the hybrid-labs-prep.ps1 script.
 
-# Participant count - change this to the number of participants you want to create
-$participantCount = 1
-
 # Participant prefix - used for the user account and resource group names
-$participantPrefix = "hybrid-"
+$participantPrefix = "<Fill this in>"
 
 # Entra ID directory name - need this for the UPN
 # $directoryName = "<FILL THIS IN>"
@@ -21,22 +18,40 @@ if (-not (Get-Module -Name Az -ListAvailable)) {
 az config set core.allow_broker=true
 az login --scope https://graph.microsoft.com//.default
 
+# Resource Graph extension is required for the query
+az extension add -n resource-graph
+
 # Set the subscription, can remove this bit for the Azure Pass deployment
 # az account set -n "<FILL THIS IN IF YOU GOT MULTIPLE SUBS>"
 
 # Get the current subscription ID
 #$subscriptionId = az account show --query id --output tsv
 
-# Loop through the number of participants, creating user accounts and resource groups, and then the ARM deployment
-for ($i = 1; $i -le $participantCount; $i++) {
-    $userNumber = "{0:D2}" -f $i
-    $participantName = $participantPrefix + $userNumber
-    $userPrincipalName = "$participantName@$directoryName"
-    $resourceGroupName = $participantName + "-rg"
+# Use Azure Resource Graph to count the number of resource groups in the subscription
+$existingResourceGroupCount = az graph query -q "resourcecontainers| where type == 'microsoft.resources/subscriptions/resourcegroups' and name startswith '$participantPrefix' | count" --query count
 
-    # Delete the resource group
-    az group delete --name $resourceGroupName --yes --no-wait
+# convert to an integer
+$existingResourceGroupCount = [int]$existingResourceGroupCount
 
-    # Delete the user account
-    az ad user delete --upn-or-object-id $userPrincipalName
+# If count is not an integer or is 0, exit
+if (-not $existingResourceGroupCount -or $existingResourceGroupCount -eq 0) {
+    Write-Host "No resource groups found with the prefix $participantPrefix"
+    exit
+} else {
+    Write-Host "Found $existingResourceGroupCount existing resource groups with the prefix $participantPrefix"
+
+    # Loop through the number of participants, creating user accounts and resource groups, and then the ARM deployment
+    for ($i = 1; $i -le $participantCount; $i++) {
+        $userNumber = "{0:D2}" -f $i
+        $participantName = $participantPrefix + $userNumber
+        $userPrincipalName = "$participantName@$directoryName"
+        $resourceGroupName = $participantName + "-rg"
+
+        # Delete the resource group
+        az group delete --name $resourceGroupName --yes --no-wait
+
+        # Delete the user account
+        az ad user delete --upn-or-object-id $userPrincipalName
+    }
 }
+
